Move scenarios to hooks
